Centralize WASM config and move Web platform code
Move Web platform sources and centralize browser-wasm build configuration.

- Renamed/moved EGL.cs and Emscripten.cs from the sample web project to 4_Operation/Graphic/src/Platforms/Web, changed namespaces to Alis.Core.Graphic.Platforms.Web and made the classes public for reuse across projects.
- Updated Emscripten.RequestAnimationFrameLoop to use IntPtr (public) and adjusted the Program.cs call site to cast the delegate pointer to IntPtr.
- Removed duplicate WASM properties/items from the sample project's csproj and added centralized WASM targets/properties in .config/target/alis.targets (GenerateGitattributesForBrowserWasm target, WasmExtraFilesToDeploy, NativeFileReference entries, Assets deploy mapping, PackageReference for CoroutineScheduler, and WASM-specific PropertyGroup including AOT/EmccFlags/PublishTrimmed/AllowUnsafeBlocks/Wasm build flags).

These changes centralize WASM build setup and expose platform code for shared use, reducing duplication in per-sample projects.

+++ 2_Application/Alis/sample/alis.sample/web/Program.cs
@@ -4,6 +4,7 @@ using System.Runtime.Versioning;
 using System.Threading.Tasks;
 using Alis.Core.Aspect.Data.Json;
 using Alis.Core.Graphic.OpenGL;
+using Alis.Core.Graphic.Platforms.Web;
 
 [assembly: SupportedOSPlatform("browser")]
 
@@ -157,7 +158,7 @@ namespace Alis.Sample.Web
 
             unsafe
             {
-                Emscripten.RequestAnimationFrameLoop((delegate* unmanaged<double, nint, int>)&Frame, nint.Zero);
+                Emscripten.RequestAnimationFrameLoop((IntPtr)(delegate* unmanaged<double, nint, int>)&Frame, nint.Zero);
             }
         }
     }
